+++ src/main/java/nextstep/theme/ThemeService.java
@@ -4,6 +4,7 @@ import static nextstep.common.exception.ExceptionMessage.INVALID_THEME_ID;
 
 import lombok.RequiredArgsConstructor;
 import nextstep.common.exception.NotExistEntityException;
+import nextstep.theme.dto.ThemeRequestDto;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -14,8 +15,8 @@ public class ThemeService {
 
     private final ThemeDao themeDao;
 
-    public Long create(ThemeRequest themeRequest) {
-        return themeDao.save(themeRequest.toEntity());
+    public Long create(ThemeRequestDto themeRequestDto) {
+        return themeDao.save(themeRequestDto.toEntity());
     }
 
     public List<Theme> findAll() {
